Fix minor typing issues
